requirements.txt
openai~=0.20.0
python-decouple
googletrans==3.1.0a0